Revert "Critical fixes for mobile browser compatibility"
This reverts commit 5c41b6a6cab1585bc587685e02d06eb3ab4d5a05.
