--- esp32_mcu/components/admin_portal/assets/debug.html
@@ -1,182 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Mobile Debug - TapGate</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 20px;
-            background-color: #f5f5f5;
-        }
-        .debug-section {
-            background: white;
-            padding: 15px;
-            margin: 10px 0;
-            border-radius: 5px;
-            border: 1px solid #ddd;
-        }
-        .debug-log {
-            background: #f9f9f9;
-            border: 1px solid #ccc;
-            padding: 10px;
-            margin: 10px 0;
-            white-space: pre-wrap;
-            font-family: monospace;
-            max-height: 300px;
-            overflow-y: auto;
-        }
-        button {
-            background: #007bff;
-            color: white;
-            padding: 10px 20px;
-            border: none;
-            border-radius: 4px;
-            cursor: pointer;
-            margin: 5px;
-        }
-        button:hover {
-            background: #0056b3;
-        }
-        .info-row {
-            margin: 5px 0;
-            padding: 5px;
-            background: #f8f9fa;
-            border-left: 3px solid #007bff;
-        }
-    </style>
-</head>
-<body>
-    <h1>Mobile Browser Debug</h1>
-    <p>This page helps diagnose mobile browser compatibility issues with TapGate.</p>
-
-    <div class="debug-section">
-        <h2>Browser Information</h2>
-        <div class="info-row">User Agent: <span id="userAgent"></span></div>
-        <div class="info-row">Current URL: <span id="currentUrl"></span></div>
-        <div class="info-row">Cookies Support: <span id="cookieSupport"></span></div>
-        <div class="info-row">JavaScript Enabled: <span style="color: green;">YES</span></div>
-        <div class="info-row">Screen Size: <span id="screenSize"></span></div>
-        <div class="info-row">Connection Type: <span id="connectionType"></span></div>
-    </div>
-
-    <div class="debug-section">
-        <h2>Network Tests</h2>
-        <button onclick="testNetworkConnectivity()">Test Network</button>
-        <button onclick="testCookies()">Test Cookies</button>
-        <button onclick="testFormSubmit()">Test Form Submit</button>
-        <button onclick="clearLogs()">Clear Logs</button>
-        <div id="debugLog" class="debug-log"></div>
-    </div>
-
-    <div class="debug-section">
-        <h2>Quick Actions</h2>
-        <button onclick="window.location.href='/enroll/'">Go to Enroll</button>
-        <button onclick="window.location.href='/main/'">Go to Main</button>
-        <button onclick="window.location.reload()">Reload Page</button>
-    </div>
-
-    <script>
-        function log(message) {
-            const debugLog = document.getElementById('debugLog');
-            const timestamp = new Date().toLocaleTimeString();
-            debugLog.textContent += `[${timestamp}] ${message}\n`;
-            debugLog.scrollTop = debugLog.scrollHeight;
-        }
-
-        function updateBrowserInfo() {
-            document.getElementById('userAgent').textContent = navigator.userAgent;
-            document.getElementById('currentUrl').textContent = window.location.href;
-            document.getElementById('cookieSupport').textContent = navigator.cookieEnabled ? 'YES' : 'NO';
-            document.getElementById('screenSize').textContent = `${screen.width}x${screen.height}`;
-            
-            // Try to detect connection type
-            let connectionType = 'Unknown';
-            if (navigator.connection) {
-                connectionType = navigator.connection.effectiveType || navigator.connection.type || 'Unknown';
-            }
-            document.getElementById('connectionType').textContent = connectionType;
-        }
-
-        function testNetworkConnectivity() {
-            log('Testing network connectivity...');
-            fetch('/', {
-                method: 'GET',
-                credentials: 'include'
-            })
-            .then(response => {
-                log(`Network test: ${response.status} ${response.statusText}`);
-                log(`Response headers: ${JSON.stringify(Object.fromEntries(response.headers.entries()))}`);
-            })
-            .catch(error => {
-                log(`Network test failed: ${error.message}`);
-            });
-        }
-
-        function testCookies() {
-            log('Testing cookie functionality...');
-            
-            // Test setting a cookie
-            const testValue = 'test_' + Date.now();
-            document.cookie = `test_cookie=${testValue}; path=/; max-age=60`;
-            
-            // Test reading the cookie
-            const cookies = document.cookie;
-            log(`Current cookies: ${cookies}`);
-            
-            if (cookies.includes(testValue)) {
-                log('Cookie test: PASSED - Can set and read cookies');
-            } else {
-                log('Cookie test: FAILED - Cannot set or read cookies');
-            }
-        }
-
-        function testFormSubmit() {
-            log('Testing form submission...');
-            
-            const formData = new FormData();
-            formData.append('test', 'mobile_debug');
-            
-            fetch('/api/debug', {
-                method: 'POST',
-                body: formData,
-                credentials: 'include'
-            })
-            .then(response => {
-                log(`Form test: ${response.status} ${response.statusText}`);
-                if (response.redirected) {
-                    log(`Form test: Redirected to ${response.url}`);
-                }
-                return response.text();
-            })
-            .then(text => {
-                log(`Form test response: ${text.substring(0, 100)}...`);
-            })
-            .catch(error => {
-                log(`Form test failed: ${error.message}`);
-            });
-        }
-
-        function clearLogs() {
-            document.getElementById('debugLog').textContent = '';
-        }
-
-        // Initialize page
-        document.addEventListener('DOMContentLoaded', function() {
-            updateBrowserInfo();
-            log('Mobile debug page loaded');
-            
-            // Log any JavaScript errors
-            window.addEventListener('error', function(e) {
-                log(`JavaScript Error: ${e.message} at ${e.filename}:${e.lineno}`);
-            });
-            
-            // Log any unhandled promise rejections
-            window.addEventListener('unhandledrejection', function(e) {
-                log(`Unhandled Promise Rejection: ${e.reason}`);
-            });
-        });
-    </script>
-</body>
-</html>
