+++ esp32_mcu/components/admin_portal/assets/page_change_psw.html
@@ -12,7 +12,7 @@
   <main class="card">
     <img src="/assets/logo.png" alt="TapGate" class="logo">
     <h1>Change Password</h1>
-    <form data-action="changePassword" action="/api/change-password" method="POST" class="form-grid">
+    <form data-action="changePassword" class="form-grid">
       <label>
         Portal name
         <input type="text" data-bind="portal-name" readonly>
