+++ esp32_mcu/components/admin_portal/assets/page_device.html
@@ -12,7 +12,7 @@
   <main class="card">
     <img src="/assets/logo.png" alt="TapGate" class="logo">
     <h1>TapGate Device</h1>
-    <form data-action="device" action="/api/device" method="POST" class="form-grid">
+    <form data-action="device" class="form-grid">
       <label>
         Access point name
         <input type="text" name="ssid" data-bind="ssid" maxlength="32" autocomplete="off" required>
