+++ esp32_mcu/components/admin_portal/assets/page_enroll.html
@@ -16,7 +16,7 @@
       Create an administrator password<br>
       to secure the TapGate portal
     </p>
-    <form data-action="enroll" action="/api/enroll" method="POST" class="form-grid">
+    <form data-action="enroll" class="form-grid">
       <label>
         Portal name
         <input type="text" name="portal" data-bind="portal-name" required placeholder="Access Point name">
