+++ esp32_mcu/components/admin_portal/http_service.c
@@ -13,9 +13,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <sys/socket.h>
-#include <arpa/inet.h>
-#include <errno.h>
 
 #include "pages/page_auth.h"
 #include "pages/page_busy.h"
@@ -123,64 +120,6 @@ static const char *page_name(admin_portal_page_t page)
 static uint64_t get_now_ms(void)
 {
     return esp_timer_get_time() / 1000ULL;
-}
-
-// Get client IP address from HTTP request
-static bool get_client_ip(httpd_req_t *req, char *ip_buffer, size_t buffer_size)
-{
-    if (!req || !ip_buffer || buffer_size < 16)
-        return false;
-
-    // Try ESP-IDF specific way first - X-Forwarded-For header
-    size_t len = httpd_req_get_hdr_value_len(req, "X-Forwarded-For");
-    if (len > 0 && len < buffer_size) {
-        if (httpd_req_get_hdr_value_str(req, "X-Forwarded-For", ip_buffer, buffer_size) == ESP_OK) {
-            LOGI(TAG, "Got client IP from X-Forwarded-For: %s", ip_buffer);
-            return true;
-        }
-    }
-
-    // Try X-Real-IP header (common in proxies)
-    len = httpd_req_get_hdr_value_len(req, "X-Real-IP");
-    if (len > 0 && len < buffer_size) {
-        if (httpd_req_get_hdr_value_str(req, "X-Real-IP", ip_buffer, buffer_size) == ESP_OK) {
-            LOGI(TAG, "Got client IP from X-Real-IP: %s", ip_buffer);
-            return true;
-        }
-    }
-
-    // Fallback to socket method
-    int sockfd = httpd_req_to_sockfd(req);
-    if (sockfd < 0) {
-        LOGW(TAG, "Failed to get socket descriptor from request");
-        return false;
-    }
-
-    // Try both getpeername (peer address) and getsockname (local address)
-    struct sockaddr_in addr;
-    socklen_t addr_len = sizeof(addr);
-    
-    // First try getpeername (remote peer address)
-    if (getpeername(sockfd, (struct sockaddr*)&addr, &addr_len) == 0) {
-        if (inet_ntop(AF_INET, &addr.sin_addr, ip_buffer, buffer_size) != NULL) {
-            // Validate the IP is not 0.0.0.0 or localhost
-            if (strcmp(ip_buffer, "0.0.0.0") != 0 && strcmp(ip_buffer, "127.0.0.1") != 0) {
-                LOGI(TAG, "Got client IP from getpeername: %s", ip_buffer);
-                return true;
-            }
-        }
-    }
-    
-    // ESP32 AP mode often can't detect client IP properly via socket calls
-    // For single-user admin portal, use a consistent mobile client identifier
-    LOGW(TAG, "Socket-based IP detection failed, using mobile client identifier");
-    
-    // For mobile browsers connecting to ESP32 AP, use a consistent identifier
-    // This ensures session consistency across requests
-    strncpy(ip_buffer, "10.10.0.100", buffer_size - 1);  // Fixed mobile client ID
-    ip_buffer[buffer_size - 1] = '\0';
-    LOGI(TAG, "Using fixed mobile client identifier: %s", ip_buffer);
-    return true;
 }
 
 static uint64_t minutes_to_ms(uint32_t minutes)
@@ -363,27 +302,9 @@ static bool get_session_token(httpd_req_t *req, char *token, size_t token_size)
         return false;
     }
 
-    // Clear buffer to prevent corruption
-    memset(buffer, 0, length + 1);
-
     if (httpd_req_get_hdr_value_str(req, "Cookie", buffer, length + 1) != ESP_OK)
     {
         LOGI(TAG, "Failed to get Cookie header value");
-        free(buffer);
-        return false;
-    }
-
-    // Validate cookie content is printable (safety check)
-    bool is_valid = true;
-    for (size_t i = 0; i < length; i++) {
-        if (buffer[i] != 0 && (buffer[i] < 32 || buffer[i] > 126)) {
-            LOGW(TAG, "Cookie contains non-printable character at position %zu (0x%02X), rejecting", i, (unsigned char)buffer[i]);
-            is_valid = false;
-            break;
-        }
-    }
-    
-    if (!is_valid) {
         free(buffer);
         return false;
     }
@@ -518,16 +439,11 @@ static esp_err_t send_redirect(httpd_req_t *req, admin_portal_page_t page)
     httpd_resp_set_status(req, "302 Found");
     httpd_resp_set_hdr(req, "Location", location);
     
-    // Add CORS headers for mobile compatibility (allow JavaScript to read Location header)
-    httpd_resp_set_hdr(req, "Access-Control-Allow-Origin", "*");
+    // Add CORS headers for mobile compatibility
+    httpd_resp_set_hdr(req, "Access-Control-Allow-Origin", "http://10.10.0.1");
     httpd_resp_set_hdr(req, "Access-Control-Allow-Credentials", "true");
-    httpd_resp_set_hdr(req, "Access-Control-Expose-Headers", "Location, Set-Cookie");
-    httpd_resp_set_hdr(req, "Access-Control-Allow-Methods", "GET, POST, OPTIONS");
-    httpd_resp_set_hdr(req, "Access-Control-Allow-Headers", "Content-Type, Cookie, Authorization");
     
     set_cache_headers(req);
-    
-    LOGI(TAG, "Redirect response headers set, sending empty body");
     return httpd_resp_send(req, "", 0);
 }
 
@@ -594,56 +510,6 @@ static admin_portal_session_status_t evaluate_session(httpd_req_t *req,
     return status;
 }
 
-// Simple session evaluation (simplified for single-client admin portal)
-static admin_portal_session_status_t evaluate_session_by_ip(httpd_req_t *req)
-{
-    char client_ip[16];
-    bool has_ip = get_client_ip(req, client_ip, sizeof(client_ip));
-    uint64_t now = get_now_ms();
-    
-    LOGI(TAG, "Evaluating simple session: client_ip=%s, now=%" PRIu64 "ms", 
-         has_ip ? client_ip : "unknown", now);
-    
-    // For admin portal, we allow only single session, so just check if any session exists
-    admin_portal_session_status_t status;
-    
-    if (has_ip && strlen(client_ip) > 0 && strcmp(client_ip, "0.0.0.0") != 0) {
-        // Use IP-based check if we have valid IP
-        status = admin_portal_state_check_session_by_ip(&g_state, client_ip, now);
-    } else {
-        // Fallback: just check if any session is active (admin portal = single user)
-        if (g_state.session.active) {
-            // Check timeout
-            if (g_state.inactivity_timeout_ms > 0) {
-                uint64_t elapsed = now - g_state.session.last_activity_ms;
-                if (elapsed > g_state.inactivity_timeout_ms) {
-                    admin_portal_state_clear_session(&g_state);
-                    status = ADMIN_PORTAL_SESSION_EXPIRED;
-                } else {
-                    status = ADMIN_PORTAL_SESSION_MATCH;
-                }
-            } else {
-                status = ADMIN_PORTAL_SESSION_MATCH;
-            }
-        } else {
-            status = ADMIN_PORTAL_SESSION_NONE;
-        }
-        LOGI(TAG, "Using session fallback (no valid IP), session active=%s", 
-             g_state.session.active ? "yes" : "no");
-    }
-    
-    LOGI(TAG, "Simple session evaluation result: status=%s", session_status_name(status));
-    
-    if (status == ADMIN_PORTAL_SESSION_MATCH)
-    {
-        admin_portal_state_update_session(&g_state, now);
-        LOGI(TAG, "Session updated with current timestamp");
-    }
-    
-    return status;
-}
-
-// Create session using hybrid approach (IP-based + cookie fallback)
 static esp_err_t ensure_session_claim(httpd_req_t *req,
                                       admin_portal_session_status_t *status,
                                       char *token_buffer,
@@ -658,31 +524,16 @@ static esp_err_t ensure_session_claim(httpd_req_t *req,
         return ESP_OK;
     }
 
-    LOGI(TAG, "Creating new hybrid session claim (IP + cookie)");
-    
-    // Try to get client IP for IP-based session
-    char client_ip[16];
-    bool has_ip = get_client_ip(req, client_ip, sizeof(client_ip));
-    
-    uint64_t now = get_now_ms();
-    
-    if (has_ip)
-    {
-        // Create IP-based session (mobile-friendly)
-        admin_portal_state_start_session_by_ip(&g_state, client_ip, now, false);
-        LOGI(TAG, "Created IP-based session for client: %s", client_ip);
-    }
-    
-    // Also create cookie-based session for desktop browsers
+    LOGI(TAG, "Creating new session claim");
     char new_token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1];
     generate_session_token(new_token);
-    LOGI(TAG, "Generated session token: %.8s...", new_token);
+    LOGI(TAG, "Generated new session token: %.8s...", new_token);
     
-    admin_portal_state_start_session(&g_state, new_token, now, false);
+    admin_portal_state_start_session(&g_state, new_token, get_now_ms(), false);
     uint32_t max_age = g_state.inactivity_timeout_ms ? (uint32_t)(g_state.inactivity_timeout_ms / 1000UL) : 60U;
     
     // For API endpoints, cookie will be set after authorization to avoid duplicate headers
-    // For GET requests (pages), set cookie immediately for desktop compatibility
+    // For GET requests (pages), set cookie immediately
     bool is_api_request = req && req->uri[0] != '\0' && strstr(req->uri, "/api/") != NULL;
     
     if (!is_api_request)
@@ -696,15 +547,14 @@ static esp_err_t ensure_session_claim(httpd_req_t *req,
     }
     
     LOGI(TAG,
-         "Hybrid session started for URI %s (IP=%s, max_age=%" PRIu32 " s)",
+         "Session started for URI %s (max_age=%" PRIu32 " s)",
          req ? req->uri : "<null>",
-         has_ip ? client_ip : "unavailable",
          max_age);
     
     strncpy(token_buffer, new_token, token_size - 1);
     token_buffer[token_size - 1] = '\0';
     *status = ADMIN_PORTAL_SESSION_MATCH;
-    LOGI(TAG, "Hybrid session claim completed successfully");
+    LOGI(TAG, "Session claim completed successfully");
     return ESP_OK;
 }
 
@@ -867,61 +717,36 @@ static esp_err_t handle_session_info(httpd_req_t *req)
 
 static esp_err_t handle_enroll(httpd_req_t *req)
 {
-    // Use IP-based session evaluation with cookie fallback
-    admin_portal_session_status_t status = evaluate_session_by_ip(req);
-    
-    // If IP-based fails, try cookie-based for desktop browsers
-    if (status == ADMIN_PORTAL_SESSION_NONE)
-    {
-        char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1] = {0};
-        status = evaluate_session(req, token, sizeof(token));
-    }
+    char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1] = {0};
+    admin_portal_session_status_t status = evaluate_session(req, token, sizeof(token));
 
     LOGI(TAG,
          "API /api/enroll: status=%s has_password=%s",
          session_status_name(status),
          admin_portal_state_has_password(&g_state) ? "true" : "false");
 
-    // Handle busy portal
-    if (status == ADMIN_PORTAL_SESSION_BUSY)
+    if (status != ADMIN_PORTAL_SESSION_MATCH)
     {
-        LOGI(TAG, "Enrollment denied: portal busy");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_BUSY);
-    }
-
-    // Start simple session if needed (with IP if available, fallback to cookie-only)
-    if (status == ADMIN_PORTAL_SESSION_NONE)
-    {
-        char client_ip[16];
-        bool has_ip = get_client_ip(req, client_ip, sizeof(client_ip));
-        
-        uint64_t now = get_now_ms();
-        
-        if (has_ip && strlen(client_ip) > 0 && strcmp(client_ip, "0.0.0.0") != 0) {
-            admin_portal_state_start_session_by_ip(&g_state, client_ip, now, false);
-            LOGI(TAG, "Started IP-based session for client: %s", client_ip);
-        } else {
-            // Fallback: start simple session without IP
-            char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1];
-            generate_session_token(token);
-            admin_portal_state_start_session(&g_state, token, now, false);
-            LOGI(TAG, "Started simple session (no valid IP)");
+        if (status == ADMIN_PORTAL_SESSION_NONE)
+            ensure_session_claim(req, &status, token, sizeof(token));
+        if (status == ADMIN_PORTAL_SESSION_BUSY)
+        {
+            LOGI(TAG, "Enrollment denied: portal busy");
+            return send_json(req, "409 Conflict", "{\"status\":\"busy\"}");
         }
     }
 
-    // Redirect to auth if password already exists
     if (admin_portal_state_has_password(&g_state))
     {
         LOGI(TAG, "Enrollment redirected to auth because password already set");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_AUTH);
+        return send_json(req, "200 OK", "{\"status\":\"redirect\",\"redirect\":\"/auth/\"}");
     }
 
     char *body = read_body(req);
     if (!body)
     {
         LOGI(TAG, "Enrollment failed: request body missing");
-        // Return to enroll page with error - could be handled by query param
-        return send_redirect(req, ADMIN_PORTAL_PAGE_ENROLL);
+        return send_json(req, "400 Bad Request", "{\"status\":\"error\",\"code\":\"invalid_request\"}");
     }
 
     char portal_name[sizeof(g_state.ap_ssid)];
@@ -942,7 +767,7 @@ static esp_err_t handle_enroll(httpd_req_t *req)
     if (!has_portal_name || portal_name[0] == '\0')
     {
         LOGI(TAG, "Enrollment failed: portal name missing");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_ENROLL);
+        return send_json(req, "200 OK", "{\"status\":\"error\",\"code\":\"invalid_ssid\"}");
     }
 
     if (!has_password || !admin_portal_state_password_valid(&g_state, password))
@@ -950,21 +775,21 @@ static esp_err_t handle_enroll(httpd_req_t *req)
         LOGI(TAG,
              "Enrollment failed: %s",
              has_password ? "password does not meet requirements" : "password missing");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_ENROLL);
+        return send_json(req, "200 OK", "{\"status\":\"error\",\"code\":\"invalid_password\"}");
     }
 
     esp_err_t err = nvm_write_string(NVM_WIFI_PARTITION, ADMIN_PORTAL_NAMESPACE, ADMIN_PORTAL_KEY_SSID, portal_name);
     if (err != ESP_OK)
     {
         LOGI(TAG, "Enrollment failed: unable to store AP SSID: %s", esp_err_to_name(err));
-        return send_redirect(req, ADMIN_PORTAL_PAGE_ENROLL);
+        return send_json(req, "500 Internal Server Error", "{\"status\":\"error\",\"code\":\"storage_failed\"}");
     }
 
     err = nvm_write_string(NVM_WIFI_PARTITION, ADMIN_PORTAL_NAMESPACE, ADMIN_PORTAL_KEY_PSW, password);
     if (err != ESP_OK)
     {
         LOGI(TAG, "Enrollment failed: unable to store password: %s", esp_err_to_name(err));
-        return send_redirect(req, ADMIN_PORTAL_PAGE_ENROLL);
+        return send_json(req, "500 Internal Server Error", "{\"status\":\"error\",\"code\":\"storage_failed\"}");
     }
 
     admin_portal_state_set_ssid(&g_state, portal_name);
@@ -978,28 +803,26 @@ static esp_err_t handle_enroll(httpd_req_t *req)
     
     admin_portal_device_set_ap_password(password);
 
-    // Set session cookie for desktop browsers (mobile will use IP-based session)
-    char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1];
-    generate_session_token(token);
+    // Set session cookie BEFORE sending response to ensure authorized session is maintained
     uint32_t max_age = g_state.inactivity_timeout_ms ? (uint32_t)(g_state.inactivity_timeout_ms / 1000UL) : 60U;
+    LOGI(TAG, "About to set session cookie for token: %.8s...", token);
     set_session_cookie(req, token, max_age);
 
     LOGI(TAG, "Enrollment successful, redirecting to main page (AP SSID=\"%s\")", portal_name);
     
-    return send_redirect(req, ADMIN_PORTAL_PAGE_MAIN);
+    // Include cookie value in JSON response for manual setting
+    char response[256];
+    snprintf(response, sizeof(response), 
+             "{\"status\":\"ok\",\"redirect\":\"/main/\",\"session_token\":\"%s\"}", 
+             token);
+    
+    return send_json(req, "200 OK", response);
 }
 
 static esp_err_t handle_login(httpd_req_t *req)
 {
-    // Use IP-based session evaluation with cookie fallback
-    admin_portal_session_status_t status = evaluate_session_by_ip(req);
-    
-    // If IP-based fails, try cookie-based for desktop browsers
-    if (status == ADMIN_PORTAL_SESSION_NONE)
-    {
-        char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1] = {0};
-        status = evaluate_session(req, token, sizeof(token));
-    }
+    char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1] = {0};
+    admin_portal_session_status_t status = evaluate_session(req, token, sizeof(token));
 
     LOGI(TAG,
          "API /api/login: status=%s has_password=%s authorized=%s",
@@ -1010,40 +833,26 @@ static esp_err_t handle_login(httpd_req_t *req)
     if (status == ADMIN_PORTAL_SESSION_BUSY)
     {
         LOGI(TAG, "Login denied: portal busy");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_BUSY);
+        return send_json(req, "409 Conflict", "{\"status\":\"busy\"}");
     }
 
-    // Start simple session if needed (with IP if available, fallback to cookie-only)
     if (status == ADMIN_PORTAL_SESSION_NONE)
     {
-        char client_ip[16];
-        bool has_ip = get_client_ip(req, client_ip, sizeof(client_ip));
-        
-        uint64_t now = get_now_ms();
-        
-        if (has_ip && strlen(client_ip) > 0 && strcmp(client_ip, "0.0.0.0") != 0) {
-            admin_portal_state_start_session_by_ip(&g_state, client_ip, now, false);
-            LOGI(TAG, "Started IP-based session for login client: %s", client_ip);
-        } else {
-            // Fallback: start simple session without IP
-            char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1];
-            generate_session_token(token);
-            admin_portal_state_start_session(&g_state, token, now, false);
-            LOGI(TAG, "Started simple session for login (no valid IP)");
-        }
+        ensure_session_claim(req, &status, token, sizeof(token));
+        LOGI(TAG, "Login request claimed new session");
     }
 
     if (!admin_portal_state_has_password(&g_state))
     {
         LOGI(TAG, "Login redirected to enroll because password missing");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_ENROLL);
+        return send_json(req, "200 OK", "{\"status\":\"redirect\",\"redirect\":\"/enroll/\"}");
     }
 
     char *body = read_body(req);
     if (!body)
     {
         LOGI(TAG, "Login failed: request body missing");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_AUTH);
+        return send_json(req, "400 Bad Request", "{\"status\":\"error\",\"code\":\"invalid_request\"}");
     }
 
     char password[sizeof(g_state.ap_password)];
@@ -1060,7 +869,7 @@ static esp_err_t handle_login(httpd_req_t *req)
     {
         LOGI(TAG, "Login failed: %s", 
              !has_password ? "password missing" : "wrong password");
-        return send_redirect(req, ADMIN_PORTAL_PAGE_AUTH);
+        return send_json(req, "200 OK", "{\"status\":\"error\",\"code\":\"wrong_password\"}");
     }
 
     LOGI(TAG, "Authorizing session after login");
@@ -1069,15 +878,20 @@ static esp_err_t handle_login(httpd_req_t *req)
     bool is_authorized = admin_portal_state_session_authorized(&g_state);
     LOGI(TAG, "Session authorization status: %s", is_authorized ? "authorized" : "not_authorized");
     
-    // Set session cookie for desktop browsers (mobile will use IP-based session)
-    char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1];
-    generate_session_token(token);
+    // Set session cookie BEFORE sending response to ensure authorized session is maintained
     uint32_t max_age = g_state.inactivity_timeout_ms ? (uint32_t)(g_state.inactivity_timeout_ms / 1000UL) : 60U;
+    LOGI(TAG, "About to set session cookie for token: %.8s...", token);
     set_session_cookie(req, token, max_age);
     
     LOGI(TAG, "Login successful, redirecting to main page");
     
-    return send_redirect(req, ADMIN_PORTAL_PAGE_MAIN);
+    // Include cookie value in JSON response for manual setting
+    char response[256];
+    snprintf(response, sizeof(response), 
+             "{\"status\":\"ok\",\"redirect\":\"/main/\",\"session_token\":\"%s\"}", 
+             token);
+    
+    return send_json(req, "200 OK", response);
 }
 
 static esp_err_t handle_change_password(httpd_req_t *req)
@@ -1202,20 +1016,8 @@ static esp_err_t handle_page_request(httpd_req_t *req, const admin_portal_page_d
     if (!desc)
         return ESP_ERR_INVALID_ARG;
 
-    // Try IP-based session first (mobile-friendly), fallback to cookies (desktop)
-    admin_portal_session_status_t status = evaluate_session_by_ip(req);
     char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1] = {0};
-    
-    if (status == ADMIN_PORTAL_SESSION_NONE)
-    {
-        status = evaluate_session(req, token, sizeof(token));
-        LOGI(TAG, "IP-based session failed, trying cookie-based: %s", session_status_name(status));
-    }
-    else
-    {
-        LOGI(TAG, "Using IP-based session: %s", session_status_name(status));
-    }
-    
+    admin_portal_session_status_t status = evaluate_session(req, token, sizeof(token));
     admin_portal_page_t target = admin_portal_state_resolve_page(&g_state, desc->page, status);
     LOGI(TAG, "Handle page request resolve page: %s", admin_portal_page_to_str(target));
 
@@ -1255,21 +1057,8 @@ static esp_err_t handle_page_request(httpd_req_t *req, const admin_portal_page_d
 static esp_err_t handle_root(httpd_req_t *req)
 {
     const admin_portal_page_descriptor_t *main_page = &admin_portal_page_main;
-    
-    // Try IP-based session first (mobile-friendly), fallback to cookies (desktop)
-    admin_portal_session_status_t status = evaluate_session_by_ip(req);
     char token[ADMIN_PORTAL_TOKEN_MAX_LEN + 1] = {0};
-    
-    if (status == ADMIN_PORTAL_SESSION_NONE)
-    {
-        status = evaluate_session(req, token, sizeof(token));
-        LOGI(TAG, "IP-based session failed for root, trying cookie-based: %s", session_status_name(status));
-    }
-    else
-    {
-        LOGI(TAG, "Using IP-based session for root: %s", session_status_name(status));
-    }
-    
+    admin_portal_session_status_t status = evaluate_session(req, token, sizeof(token));
     admin_portal_page_t target = admin_portal_state_resolve_page(&g_state, main_page->page, status);
     LOGI(TAG, "Handle root resolve page: %s", admin_portal_page_to_str(target));
